feat: use shared chart utils in hog-charts
Import formatValue, createXAxisTickCallback, buildTheme, and seriesColor
from lib/charts/utils/ instead of duplicating them. Deletes 3 dead util
files and makes HogChartTheme extend the shared ChartTheme.
