--- frontend/src/lib/hog-charts/builders/index.ts
@@ -1,8 +1,0 @@
-export { buildGoalLines, buildLineSeries, buildYAxis } from './line'
-export type {
-    BuildLineSeriesOptions,
-    BuildLineSeriesResult,
-    DataSeries,
-    GoalLineInput,
-    StatisticalOverlays,
-} from './line'
